frontend/requirements.txt
streamlit>=1.29.0
Pillow>=10.2.0